+++ src/wai/annotations/subdir/ac/specifier/_SubDirOutputFormatSpecifier.py
@@ -1,0 +1,5 @@
+from wai.annotations.domain.audio.classification import AudioClassificationDomainSpecifier
+
+from ...common.specifier import output_format_specifier
+
+SubDirOutputFormatSpecifier = output_format_specifier(AudioClassificationDomainSpecifier, "audio files")