+++ src/wai/annotations/subdir/common/component/__init__.py
@@ -1,4 +1,4 @@
 from ._SubDirACReader import SubDirACReader
 from ._SubDirACWriter import SubDirACWriter
-from ._SubDirReader import SubDirReader
+from ._SubDirReader import reader_class
 from ._SubDirWriter import SubDirWriter