+++ src/wai/annotations/subdir/common/specifier/__init__.py
@@ -1,4 +1,4 @@
 from ._SubDirACInputFormatSpecifier import SubDirACInputFormatSpecifier
 from ._SubDirACOutputFormatSpecifier import SubDirACOutputFormatSpecifier
-from ._SubDirInputFormatSpecifier import SubDirInputFormatSpecifier
-from ._SubDirOutputFormatSpecifier import SubDirOutputFormatSpecifier
+from ._SubDirInputFormatSpecifier import input_format_specifier
+from ._SubDirOutputFormatSpecifier import output_format_specifier