+++ CHANGES.rst
@@ -1,11 +1,16 @@
 Changelog
 =========
 
+2.0.0 (2021-12-16)
+------------------
+
+- Added support for spectrum classification domain: `from-subdir-sc` and `to-subdir-sc`.
+- Split into base/specialisation structure to allow for easy specification of classification domains.
+
 1.0.1 (2022-06-21)
 ------------------
 
 - added reader/writer for audio classification: `from-subdir-ac` and `to-subdir-ac`
-
 
 1.0.0 (2021-05-20)
 ------------------